types: fold the membership tests into TaskSet.test.ts
One suite per module — the file converts to @effect/vitest so the
existing sync tests and the database-backed membership suites coexist.
